完善消费者，完善个人信息接口和搜索接口

+++ user-service/user-service-server/src/main/java/com/gitgle/consumer/message/TalentRankMessage.java
@@ -10,6 +10,6 @@ public class TalentRankMessage implements Serializable {
 
     private String login;
 
-    private BigDecimal talentRank;
+    private Double talentRank;
 
 }
